refactor(server): rename route modules to route.ts within their folders
The internal/external folders already namespace these, so processRoute.ts ->
internal/route.ts and completionsRoute.ts -> external/route.ts. Update the two
imports and doc-comment references in server.ts. No behavior change.
