+++ apps/client/server/chatCompletion/external/route.ts
