+++ apps/client/server/chatCompletion/internal/route.ts
